Refactor notification domain: Rename packages and introduce NotificationAction, NotificationEvent, NotificationState, and NotificationViewModel
