+++ app/src/main/java/com/notifier/app/notification/domain/model/Owner.kt
@@ -1,4 +1,4 @@
-package com.notifier.app.notification.domain
+package com.notifier.app.notification.domain.model
 
 data class Owner(
     val avatarUrl: String,
