+++ app/src/main/java/com/notifier/app/notification/presentation/NotificationEvent.kt
@@ -1,0 +1,9 @@
+package com.notifier.app.notification.presentation
+
+import com.notifier.app.core.domain.util.NetworkError
+import com.notifier.app.core.domain.util.PersistenceError
+
+sealed interface NotificationEvent {
+    data class NetworkErrorEvent(val error: NetworkError) : NotificationEvent
+    data class PersistenceErrorEvent(val error: PersistenceError) : NotificationEvent
+}
